attach pricing context to system metrics
